added vaf tooltip and used it in  snv mutation arc
