+++ client/plots/disco/snv/NonExonicSnvArcsMapper.ts
@@ -43,6 +43,8 @@ export default class NonExonicSnvArcsMapper {
 				mname: data.mname,
 				chr: data.chr,
 				pos: data.position,
+				refCount: data.refCount,
+				altCount: data.altCount,
 				sampleName: [data.sampleName]
 			}
