+++ client/plots/disco/snv/SnvArcsMapper.ts
@@ -58,6 +58,8 @@ export default class SnvArcsMapper {
 						mname: data.mname,
 						chr: data.chr,
 						pos: data.position,
+						refCount: data.refCount,
+						altCount: data.altCount,
 						sampleName: [data.sampleName]
 					}
 					snvArray.push(arc)
